Modularised the command and the function, and stopped further reading when word is found
