+++ commands/translate.js
@@ -1,7 +1,4 @@
-const fs = require('fs');
-const readline = require('readline');
-const { appFolder } = require('../constants');
-const readStream = fs.createReadStream([appFolder, 'file.dat'].join("/"));
+const { translateToGaro } = require('../helpers');
 
 exports.command = 'translate [key]';
 exports.describe = 'Get the Garo Translation of the word';
@@ -11,15 +8,5 @@ exports.handler = (argv) => {
     console.error('Please provide a word to translate');
   }
 
-  const rl = readline.createInterface({
-    input: readStream,
-    crlfDelay: Infinity,
-  });
-  
-  rl.on('line', (line) => {
-    [english, garo] = line.split('=');
-    if( english === key ) {
-      console.log(`${english}: ${garo}`);
-    }
-  });
+  translateToGaro(key);
 };
